renames application folder files and packages to asset
