+++ cmd/flawfix/main.go
@@ -26,7 +26,7 @@ import (
 	"github.com/l3montree-dev/flawfix/internal/auth"
 	"github.com/l3montree-dev/flawfix/internal/core"
 
-	"github.com/l3montree-dev/flawfix/internal/core/application"
+	"github.com/l3montree-dev/flawfix/internal/core/asset"
 	"github.com/l3montree-dev/flawfix/internal/core/env"
 	"github.com/l3montree-dev/flawfix/internal/core/flaw"
 	"github.com/l3montree-dev/flawfix/internal/core/org"
@@ -91,7 +91,7 @@ func main() {
 	// we need those core features in globally scoped middlewares. Therefore
 	// initialize them right here.
 	patRepository := pat.NewGormRepository(db)
-	appRepository := application.NewGormRepository(db)
+	assetRepository := asset.NewGormRepository(db)
 	projectRepository := project.NewGormRepository(db)
 	projectScopedRBAC := project.ProjectAccessControlFactory(projectRepository)
 
@@ -114,9 +114,9 @@ func main() {
 	// since this application has a multi tenant and hierarchical structure
 	// we need to pass the returned router to the next registration.
 	tenantRouter := org.RegisterHttpHandler(db, sessionRouter, casbinRBACProvider)
-	projectRouter := project.RegisterHttpHandler(db, tenantRouter, appRepository)
-	applicationRouter := application.RegisterHttpHandler(db, projectRouter, projectScopedRBAC)
-	envRouter := env.RegisterHttpHandler(db, applicationRouter, appRepository)
+	projectRouter := project.RegisterHttpHandler(db, tenantRouter, assetRepository)
+	applicationRouter := asset.RegisterHttpHandler(db, projectRouter, projectScopedRBAC)
+	envRouter := env.RegisterHttpHandler(db, applicationRouter, assetRepository)
 
 	flaw.RegisterHttpHandler(db, envRouter, projectScopedRBAC)
