+++ internal/core/asset/asset_controller.go
@@ -1,4 +1,4 @@
-package application
+package asset
 
 import (
 	"github.com/google/uuid"
