+++ internal/core/asset/asset_model.go
@@ -1,4 +1,4 @@
-package application
+package asset
 
 import (
 	"github.com/google/uuid"
