+++ internal/core/asset/asset_repository.go
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package application
+package asset
 
 import (
 	"github.com/google/uuid"
