+++ internal/core/project/project_handler.go
@@ -3,19 +3,19 @@ package project
 import (
 	"github.com/l3montree-dev/flawfix/internal/accesscontrol"
 	"github.com/l3montree-dev/flawfix/internal/core"
-	"github.com/l3montree-dev/flawfix/internal/core/application"
+	"github.com/l3montree-dev/flawfix/internal/core/asset"
 )
 
 func RegisterHttpHandler(
 	database core.DB,
 	server core.Server,
-	appRepository application.Repository,
+	assetRepository asset.Repository,
 ) core.Server {
 	database.AutoMigrate(&Model{})
 
 	repository := NewGormRepository(database)
 
-	controller := NewHttpController(repository, appRepository)
+	controller := NewHttpController(repository, assetRepository)
 
 	server.GET("/projects/", controller.List, core.AccessControlMiddleware("organization", accesscontrol.ActionRead))
